bootil: patch for aarch64 support
This package has been disabled on aarch64-linux since #51813 in 2018, but was
not updated to include aarch64-darwin in badPlatforms, so it's been failing
there. While investigating the build failure, I found that it would be trivial
to patch bootil to compile on aarch64.

The only x86-specific code in bootil is in src/3rdParty/smhasher/Platform.h,
which defines a function which references the x86 RDTSC instruction using
inline assembly. That function is neither exported nor used by the library
itself, so I've patched it out and re-enabled aarch64 compilation.