recoll: use wrapProgram to set $PATH for recoll, recollindex
Many recoll filters assume that $PATH contains all their
dependencies.  Our current patching approach is a game of
whack-a-mole, and while we will continue trying to play it we ought
to set the $PATH as well.

This commit eliminates a huge number of extraneous error messages
from recoll's log, making it easier to see the remaining ones which
really matter.