firefox: support unsigned extensions in wrapper
A similar technique is done by [Arch][1], [Debian][2], [Fedora][3],
[Gentoo][4], and probably others.

Since we are only using the `app` scope, not the `system` one, don’t use
it to minimize the expansion of allowed unsigned extension paths.

Since we can now have unsigned addons in `nixExtensions` with Firefox
release builds, remove various assertions around scenarios where they
would previously not work.

The addons *must* be in `../lib/firefox/browser/extensions` relative to
the firefox exe. So do some dance to ensure the `install_url` is correct
in the policy.

Discovering the required path by code-diving was…less than successful as
there is a lot of indirection in the handling in FF source. Instead, the
browser toolbox was used to inspect the value of
`lazy.XPIInternal.XPIStates.locations()`.

See comments in code for some more elaboration.

[1]: https://gitlab.archlinux.org/archlinux/packaging/packages/firefox/-/blob/29ab50162f4a89d281b9f0a0fb9e2ebce496e4ac/PKGBUILD#L120
[2]: https://salsa.debian.org/mozilla-team/firefox/-/blob/90a0107053660c575fa432615830dedf8c9b7e13/debian/browser.mozconfig.in#L30
[3]: https://src.fedoraproject.org/rpms/firefox/blob/e9c16e2628afd9e0d7f0962228bec7b0f56172d4/f/firefox-mozconfig#_16
[4]: https://gitweb.gentoo.org/repo/gentoo.git/tree/www-client/firefox/firefox-114.0.ebuild?id=0faf645c97c0d979fb10cb92e0d7e5980ffb1ea6#n815