mupdf: actually build and install the shared libraries version
The current version of the derivation builds in shared libraries mode,
but then the "shared" flag is not passed to the "make install"
invocation. This causes "make install" to perform a whole second build
from scratch, in static mode, and install this to $out instead.

Instead pass shared=yes as part of the makeFlags -- this is basically
the only thing that the "shared" build target does anyway, and I don't
think there is a similar target for "make install".

This issue was detected because the .pc shipped with mupdf doesn't work
with the static libraries currently shipped (it doesn't include
recursive dependencies like zlib).

(cherry picked from commit 3d4769a9e86a51117973281d8ab4328713c4c180)