mesa: apply patch to fix gallium-xa detection
This commit applies a patch submitted upstream that allows mesa's
meson to set `gallium-xa` automatically depending on (a large number
of) other features' enablement/disablement.

  https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/23581

In order to get mesa to do this detection for us, we must use
`mesonAutoFeatures="auto"` for mesa; otherwise it cannot distinguish
between `auto` and `enabled`.

Without this commit, mesa does not make it past the configurePhase
on powerpc64le.