python38Packages.httplib2: resurrect
Without this patch, this is what happened:

```sh
➤ nix-build -A python38Packages.httplib2
error: flask-sqlalchemy-3.0.3 not supported for interpreter python3.8
(use '--show-trace' to show detailed location information)
```

That missing dependency was just for testing, but it doesn't make much sense to remove a package just because it can't be tested.

I just changed it so that it only runs checks when running at python >= 3.9, and now httplib2 is available for Python 3.8 again.