treewide: fetchgit -> fetchzip (git.kernel.org)
fetchzip is more efficient, because it doesn't do a full git clone, so
it should be preferred where possible.

Where hashes have not been changed, I have verified that they don't
need to be.  Where hashes have changed, in all cases this is because
of .gitattributes files that exclude certain files from the tarball,
and in these cases I have verified that the packages still build.

sbsigntool still uses fetchgit because it has a submodule, and ell and
iwd still use fetchgit because git.kernel.org does not provide
snapshot links for them.  Apparently this is intentional.