code-server: 4.12.0 -> 4.13.0
I replaced the commit with the actual commit instead of "none" and fixed
the other replacement (`$commit` did not exist).

There was at least one postinstall script in `extensions` that was not
being ran so I modified the `find` command to account for lock files in
that directory in addition to `node_modules`.

Lastly, inject the version into the package.json otherwise it uses the
placeholder version 0.0.0.