python3Packages.certifi: allow not overriding ca bundle
We replace certifi's CA bundle with the system-wide bundle, however, the system-wide bundle includes the openssl-specific BEGIN TRUSTED CERTIFICATE cert format, while the bundle that ships with certifi contains only the BEGIN CERTIFICATE cert format.

This breaks some things that expect to see only the latter format, so I'm adding the ability to not override the CA bundle.