pkgs/top-level: Improve crossSystem equality early on
This is a workaround that lets us use `==` on systems despite the
functions that would make `==` always return `false` if a
workaround like this is not applied.

The real solution is to either use `lib.systems.equals` for all
such comparisons, or to remove all functions from the elaborated
systems.