ntopng: fix missing static assets
Instead of building the `dist` make target in the derivation (requires npm and
all that), we're using the upstream pre-built static assets. Upstream refers to
these using a sub-module so we need to fetch that.

Without this patch the web UI loads the html, but 404s on all CSS and JS assets
because the `$out/share/ntopng/httpdocs/dist` directory is empty