treewide: expunge functions from platforms
This commit moves `canExecute`, `emulator`, and `emulatorAvailable`
out of `{host,build,target}Platform` and lifts them into
`lib.systems`.  Stubs have been left behind to assist with migration
of out-of-tree code.  These will be changed to `lib.warn` or
`throw`, but that change will not be backported to 23.05.

Nix's function equality is unpredictable and error-prone; we should
not rely on it.  Since we need to be able to test
`{host,build,target}Platform` attrsets for equality, the most
sustainable long-term solution is to forbid functions from these
attrsets, much like we do for the `meta` attrset.