lib.systems.elaborate: make placeholder functions always identical
Comparing platform sets using `==` works for most cases since the in the
native case, the platform sets would not only be equal, but the same
identical value (in terms of memory location). This is, however,
technically not always the case, e.g. when specifying crossSystem
explicitly, cases can arise when we should use a native stdenv, but
don't since the functions wouldn't compare as equal due to them being
constructed in separate invocations of `lib.system.elaborate`. This
problem is described in #237427 and motivated #237512 as an equality
predicate that ignors the functions.

Since the functions in the platform attribute set no longer relate to
the platform set at all, we can use the same function value for every
placeholder deprecation error function in the attribute set—consequently
they will all compare as equal even for independently constructed
platform sets.

This is achieved by floating the funtions into a let binding in
`lib/systems/default.nix`, ensuring that they are only constructed once
per nixpkgs instance, i.e. when that file is imported. (Consequently,
comparison of platform sets should be reliable as long as they come from
the same instance of nixpkgs.)

    let
      inherit (
        import <nixpkgs> {
          localSystem = "x86_64-linux";
          crossSystem.system = "x86_64-linux";
        }
      ) buildPlatform
        hostPlatform
        ;
    in

    buildPlatform == hostPlatform
    # => true

This property is also verified with new test cases.