lib/systems: add and check valid kernels for each abi
Each kernel supports only a subset of the universe of possible ABIs.
This commit explicitly lists (opt-in) the allowable abis for each
kernel.  Attempting to parse a kernel with an abi which is not on
the list will fail.  This "search space pruning" is essential to
getting our gnu-config agreement tests down to a manageable set of
test cases.  It also lets us reject nonsense triples instead of
trying to handle them exactly the same way gnu-config does.