substitute-all.nix: Define in terms of the `substitute` bash function
The `substituteAll` bash function that was previously used substitutes
all environment variables. This can cause surprising behaviour, because
some environment variables, like `system`, and not under control of the
expression author that calls `substituteAll`.

This commit changes the nixpkgs `substituteAll` function to use
`substitute` with an explicit variable list instead. This ensures that
the only variables that will be substituted are those that the caller
specified, and not any that stdenv.mkDerivation or Nix itself
added/modified.

Partially addresses #237216.