lib/systems: move powerpc64-be ABI-defaulting out of the parser
In PR #182807 we decided that big-endian PowerPC64 should default to
the new elfv2 ABI.  The implementation in that PR did the defaulting
by adding two lines in the *triple parser*, which is turning out to
be problematic; it means that our parse-then-unparse roundtrip
disagrees with gnu-config.

This commit therefore reverts just those two lines.  If the
defaulting logic needs to be moved elsewhere that should be done.

https://github.com/NixOS/nixpkgs/pull/182807/files#r1234650738