lib/systems: introduce tripleFromSystemLossy
Unfortunately gnu-config triples are, in some rare cases, sensitive
to whether the vendor was specified as `unknown` or was omitted.  In
other words, there are situations where these are not handled
exactly the same way:

- `mips-linux-gnu`
- `mips-unknown-linux-gnu`

This commit introduces `lib.systems.tripleFromSystemLossy`, which
returns `"mips-unknown-linux-gnu"` for both of the platforms above.
This is almost always what you want to use.