nixos/nftables: extend check redirects
The `checkPhase` script that runs `nft --check` uses libredirect to link
/etc/services and /etc/protocols inside the sandbox.
The problem is that it resets NIX_REDIRECTS *after* running the
user-provided `preCheckRuleset`, so it's not possible to add more
redirects.

This adds the generated /etc/hosts as well, needed to resolve static
hostnames, and a `:$NIX_REDIRECTS` to make this extensible.