nixos/fontconfig: remove misleading documentation
These seemingly misguided comments were added by 65592837b6e62fb555d6e8c891f347428886c4f2,
and are the source of the confusion in #194594 and #222236.

For reference, Android ships on devices with 400-500 DPI displays, and
still makes use of greyscale anti-aliasing. It does not use subpixel AA
which *does* hit diminishing returns at high resolutions, but can cause
visible artifacts (at 300 DPI, we already have three times more pixels
than a standard density display, catching up with the effective subpixel
resolution of those screens).

For a subjective reference point, I already find subpixel anti-aliasing
ineffective on a ~100 DPI monitor, and the color fringing is noticeable
and intrusive.

According to [1], the v40 TrueType interpreter does not do subpixel
hinting. Instead, it aligns to whole pixels. This helps with sharpness,
and I have no reason to believe that it is harmful at higher
resolutions.
It is still useful on displays that are between standard (96 DPI) and
"retina" (250-300+ DPI) density. These are generally considered high-DPI
because they require scaling. For example, a 14" 1080p laptop, or a 27"
4k monitor are around 150 DPI.

[1]: https://freetype.org/freetype2/docs/hinting/subpixel-hinting.html