codeql: fix passing in nix JDK
The old approach only set CODEQL_JAVA_HOME in codeql, but apparently
that didn't propagate to other tools such as the various
`autobuild.sh` scripts. Setting it in a wrapper did seem to work
for `codeql database create testdb --language javascript` but not
for `codeql database analyze ...`.

This approach based on the earlier PR by Bas fixed both commands
for me.

Co-Authored-By: Bas Alberts <bas@anti.computer>