darwin."10.13.6".xnu: Complete versions in fake availability.pl
These are the versions from `AvailabilityVersions/availability.pl` for
macOS 10.13.6. The only reason we create the shell script is to avoid a
dependency on Perl in the bootstrap.

We should consider packaging AvailabilityVersions so the script becomes
something that consumes `availability.pl`, that way we don't need to
keep the versions synchronized manually.