scenebuilder: 15.0.1 -> 19.0.0
Bump scenebuilder to 19.0.0 as the latest release 20.0.0 depends on
jdk20, which is not packaged in nixpkgs yet.

Since version 15 the build system for scenebuilder was switched from
gradle to maven.

Additionally this commit switches to using a jre which bundles
openjfx. This used to be the default but was changed in ce6bc62,
which prevented scenebuilder from starting.