ath9k-htc-blobless-firmware: use name expected by the kernel
The kernel asks for the firmware with the version string in the filename[1], so
it's not easy to load the "blobs" from this package.
With this change you can just add the package with a `lib.hiPrio` to
`hardware.firmware` and it will be loaded correctly.

Note: I used symlinks instead of simply renaming the files in order to not
introduce a breaking change.

(cherry picked from commit cc065fe18a566677be6b2b45954be0b4cef33c61)