jetbrains: fix application icon
My Desktop doesn't recognize the icon in ~/.local/share/pixmaps which is
how the icon would have to be found when doing a user profile install.

For me, it's hard to find information on this, but it seems like
share/pixmaps is something old and the share/icons location should be
used instead:

https://www.reddit.com/r/linuxquestions/comments/htg6c/comment/c1yfi0a/

https://github.com/spyder-ide/spyder/issues/6188

https://github.com/FreeSpacenav/spnavcfg/issues/9

In this PR we _add_ the icons location so that both are available for
compatibility.