installShellFiles: ensure shell completion has more than 256 bytes
I found an error in the completion script of tmuxp which were outputting
an error message instead of a proper script so a sanity check in the
installShellCompletion script is helpful to avoid this to happen in the
future.

After looking at possible ways to detect this, the only valid check I
found was based on the generated file size as any completion script I
could find had more than 256 bytes.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>